+++ CHANGELOG.md
@@ -1,13 +1,19 @@
 # JLed changelog
 
+## [2019-06-20] 4.2.0
+
+* changing an effect resets the Jled object so it starts over with the 
+  new effect (see #25). Prior to this change, calling `Reset()` was
+  necessary.
+
 ## [2019-05-11] 4.1.2
 
 * fix: ESP32 dynamic channel assignment fixed. Sequence demo now working
-       as expected.
+       as expected (see #22).
 
 ## [2019-05-07] 4.1.1
 
-* fix: version format in library.properties (removed leading `v`)
+* fix: version format in library.properties (removed leading `v`; see #21)
 
 ## [2019-03-10] v4.1.0